+++ tests/pure/es.object.define-property.js
@@ -12,6 +12,15 @@ QUnit.test('Object.defineProperty', assert => {
   });
   assert.same(result, source);
   assert.same(result.q, 42);
+
+  if (DESCRIPTORS) {
+    // eslint-disable-next-line prefer-arrow-callback -- required for testing
+    assert.same(defineProperty(function () { /* empty */ }, 'prototype', {
+      value: 42,
+      writable: false,
+    }).prototype, 42, 'function prototype with non-writable descriptor');
+  }
+
   assert.throws(() => defineProperty(42, 1, {}));
   assert.throws(() => defineProperty({}, create(null), {}));
   assert.throws(() => defineProperty({}, 1, 1));